Lok-Kontroll-Funktionen ausgelagert
