+++ MEKB_H0_Anlage/Hauptform/Form1.cs
@@ -427,303 +427,6 @@ namespace MEKB_H0_Anlage
         #endregion
 
         #region Lok Kontrolle
-        private void LokKontroll_Name_SelectedIndexChanged(object sender, EventArgs e)
-        {
-            if (sender is ComboBox Namensfeld)
-            {
-                //Name der Instanz finden
-                string name = Namensfeld.Name;
-                string[] subs = name.Split('_');
-                if (subs[1] != "Name") return; //Muss auf _Name enden
-
-                //Ausgewählte Lok heraussuchen
-                if(LokomotivenArchiv.SucheDurchName(Namensfeld.Text,out Lokomotive lokomotive))
-                {
-                    LokKontroll_UpdateAktiveLok(subs[0], lokomotive);
-                }
-            }
-        }
-        private void LokKontroll_Adr_ValueChanged(object sender, EventArgs e)
-        {
-            if (sender is NumericUpDown Adressfeld)
-            {
-                //Name der Instanz finden
-                string name = Adressfeld.Name;
-                string[] subs = name.Split('_');
-                if (subs[1] != "Adr") return; //Muss auf _Adr enden
-              
-                if (LokomotivenArchiv.SucheDurchAdresse(Adressfeld.Value, out Lokomotive lokomotive))//Finde Lok mit dieser Adresse 
-                {
-                    LokKontroll_UpdateAktiveLok(subs[0], lokomotive);
-                }
-                else // Lok nicht gefunden
-                {
-                    lokomotive = new Lokomotive() //Blanke Lok anlegen mit dieser Adresse
-                    {
-                        Adresse = ((int)Adressfeld.Value),
-                        Name = String.Format("Lok: {0}", Adressfeld.Value)
-                    };            
-                    LokKontroll_UpdateAktiveLok(subs[0], lokomotive);
-                }
-            }
-        }
-        private void LokKontroll_Update_Rufnummern(object sender, EventArgs e)
-        {
-            if (sender is ComboBox Gattungsfeld)
-            {
-                //Name der Instanz finden
-                string name = Gattungsfeld.Name;
-                string[] subs = name.Split('_');
-                if (subs[1] != "Gattung") return; //Muss auf _Gattung enden
-
-                //Aktiver Lokindex
-                string indexStr = subs[0].Substring(7);
-                if (Int32.TryParse(indexStr, out int index))
-                {
-                    index--;
-                    if (index < 0) return;
-                    if (index > AktiveLoks.Length) return;
-                }
-                else
-                {
-                    return;
-                }
-
-                AktiveLoks[index].Gattung = Gattungsfeld.Text;
-
-                LokKontroll_UpdateAll(subs[0]);
-                
-            }
-        }
-        private void LokKontroll_OpenFahrpult_Click(object sender, EventArgs e)
-        {
-            if (sender is Button button)
-            {
-                if (!(Weichen_Init & Signal_Init))
-                {
-                    MessageBox.Show("Initialisierung der Z21 nicht abgeschlossen. Bitte warten und erneut versuchen", "Fenster kann nicht geöffnet werden", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
-                    return;
-                }
-
-                //Name der Instanz finden
-                string name = button.Name;
-                string[] subs = name.Split('_');
-                if (subs[1] != "OpenFahrpult") return; //Muss auf _Name enden
-
-                //Passendes Adressfeld heraussuchen
-                NumericUpDown Adressfeld = (NumericUpDown)this.Controls.Find(subs[0] + "_Adr", true)[0];
-                if (Adressfeld == null) return; //Nicht gefunden: Abbrechen
-
-                if (Adressfeld.Value == 0)
-                {
-                    MessageBox.Show("Keine Lok gewählt", "Fenster kann nicht geöffnet werden", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                    return;
-                }
-
-                //Aktiver Lokindex
-                string indexStr = subs[0].Substring(7);
-                if (Int32.TryParse(indexStr, out int index))
-                {
-                    index--;
-                    if (index < 0) return;
-                    if (index > AktiveLoks.Length) return;
-                }
-                else
-                {
-                    return;
-                }
-                if (!AktiveLoks[index].Steuerpult.IsDisposed)
-                {
-                    AktiveLoks[index].Steuerpult.Dispose();
-                }
-                AktiveLoks[index].Steuerpult = new ZugSteuerpult(AktiveLoks[index]);
-                AktiveLoks[index].Register_CMD_LOKFUNKTION(Setze_Lok_Funktion);
-                AktiveLoks[index].Register_CMD_LOKFAHRT(Setze_Lok_Fahrt);
-                AktiveLoks[index].Register_CMD_LOKSTATUS(Setze_Lok_Status);
-                AktiveLoks[index].Steuerpult.Show();
-                z21Start.Z21_GET_LOCO_INFO(AktiveLoks[index].Adresse);
-
-            }
-        }
-        private void LokKontroll_Strg_Typ_Click(object sender, EventArgs e)
-        {
-            if (sender is Button steuertyp)
-            {
-                //Name der Instanz finden
-                string name = steuertyp.Name;
-                string[] subs = name.Split('_');
-                if (subs[1] != "Strg") return; //Muss auf _Strg an zweiter Position haben
-
-                //Aktiver Lokindex
-                string indexStr = subs[0].Substring(7);
-                if (Int32.TryParse(indexStr, out int index))
-                {
-                    index--;
-                    if (index < 0) return;
-                    if (index > AktiveLoks.Length) return;
-                }
-                else
-                {
-                    return;
-                }
-
-                if (!AktiveLoks[index].Automatik)
-                {
-                    steuertyp.Text = "Automatik";
-                    steuertyp.BackColor = Color.FromArgb(0, 0, 255);
-                    AktiveLoks[index].Automatik = true;
-                }
-                else
-                {
-                    steuertyp.Text = "Manuell";
-                    steuertyp.BackColor = Color.FromArgb(0, 128, 0);
-                    AktiveLoks[index].Automatik = false;
-                }
-            }
-        }
-        private void LokKontroll_Stop_Click(object sender, EventArgs e)
-        {
-            if (sender is Button stop)
-            {
-                //Name der Instanz finden
-                string name = stop.Name;
-                string[] subs = name.Split('_');
-                if (subs[1] != "Stop") return; //Muss auf _Strg an zweiter Position haben
-
-                //Aktiver Lokindex
-                string indexStr = subs[0].Substring(7);
-                if (Int32.TryParse(indexStr, out int index))
-                {
-                    index--;
-                    if (index < 0) return;
-                    if (index > AktiveLoks.Length) return;
-                }
-                else
-                {
-                    return;
-                }
-                if (AktiveLoks[index].Adresse != 0)
-                {
-                    Setze_Lok_Fahrt(AktiveLoks[index].Adresse, 255, AktiveLoks[index].Richtung, AktiveLoks[index].FahrstufenInfo);
-                }
-            }
-        }
-
-        public void LokKontroll_UpdateAktiveLok(string CtrlID, Lokomotive lokomotive)
-        {
-            string indexStr = CtrlID.Substring(7);
-            if (Int32.TryParse(indexStr, out int index))
-            {
-                index--; //Start bei 0
-                if (index < 0) return;
-                if (index > AktiveLoks.Length) return;
-            }
-            else //Kein Index erkannt
-            {
-                return;
-            }
-            AktiveLoks[index] = lokomotive;
-            AktiveLoks[index].Automatik = false;
-            LokKontroll_UpdateAll(CtrlID);
-        }
-        public void LokKontroll_UpdateAll(string CtrlID)
-        {
-            NumericUpDown Adressfeld;
-            Button SuchFeld;
-            ComboBox Gattungsfeld;
-            Button Steuertypfeld;
-            TextBox Ruffeld;
-
-            try
-            {
-                //Passendes Adressfeld heraussuchen
-                Adressfeld = (NumericUpDown)this.Controls.Find(CtrlID + "_Adr", true).First();
-                if (Adressfeld == null) return; //Nicht gefunden: Abbrechen
-
-                //Passendes Steuertypfeld finden
-                SuchFeld = (Button)this.Controls.Find(CtrlID + "_Suche", true).First();
-                if (SuchFeld == null) return; //Nicht gefunden: Abbrechen
-
-                //Passendes Gattungsfeld finden
-                Gattungsfeld = (ComboBox)this.Controls.Find(CtrlID + "_Gattung", true)[0];
-                if (Gattungsfeld == null) return; //Nicht gefunden: Abbrechen
-
-                //Passendes Steuertypfeld finden
-                Steuertypfeld = (Button)this.Controls.Find(CtrlID + "_Strg_Typ", true)[0];
-                if (Steuertypfeld == null) return; //Nicht gefunden: Abbrechen
-
-                //Passendes Ruffeld finden
-                Ruffeld = (TextBox)this.Controls.Find(CtrlID + "_Ruf", true)[0];
-                if (Ruffeld == null) return; //Nicht gefunden: Abbrechen
-            }
-            catch
-            {
-                return;
-            }
-
-            string indexStr = CtrlID.Substring(7);
-            if (Int32.TryParse(indexStr, out int index))
-            {
-                index--; //Start bei 0
-                if (index < 0) return;
-                if (index > AktiveLoks.Length) return;
-            }
-            else //Kein Index erkannt
-            {
-                return;
-            }
-            
-            Adressfeld.Value = AktiveLoks[index].Adresse;
-            SuchFeld.Text = AktiveLoks[index].Name;
-            Gattungsfeld.Text = AktiveLoks[index].Gattung;
-            if (AktiveLoks[index].Adresse != 0) Ruffeld.Text = LokomotivenArchiv.Abkuerung[AktiveLoks[index].Gattung] + AktiveLoks[index].Adresse.ToString();
-            else Ruffeld.Text = "";
-
-            if (AktiveLoks[index].Automatik)
-            {
-                Steuertypfeld.Text = "Automatik";
-                Steuertypfeld.BackColor = Color.FromArgb(0, 0, 255);
-            }
-            else
-            {
-                Steuertypfeld.Text = "Manuell";
-                Steuertypfeld.BackColor = Color.FromArgb(0, 128, 0);
-            }
-        }
-
-        private void SearchLok_Click(object sender, EventArgs e)
-        {
-            if (sender is Button stop)
-            {
-                //Name der Instanz finden
-                string name = stop.Name;
-                string[] subs = name.Split('_');
-                if (subs[1] != "Suche") return; //Muss auf _Strg an zweiter Position haben
-
-                //Aktiver Lokindex
-                if (!ThreadLoksuche.IsAlive)
-                {
-                    ThreadLoksuche = new Thread(() => DialogHandhabungLokSuche(subs[0]));
-                    ThreadLoksuche.Start();
-                }
-            } 
-        }
-
-        private void DialogHandhabungLokSuche(string IndexName)
-        {
-            LokSuche lokSuche = new LokSuche(LokomotivenArchiv);
-
-            DialogResult dialogResult = lokSuche.ShowDialog();
-            if (dialogResult == DialogResult.OK)
-            {
-                this.BeginInvoke((Action<string, Lokomotive>)SchreibeSuchergebnis, IndexName, lokSuche.GewaehlteLok);
-            }
-        }
-
-        private void SchreibeSuchergebnis(string IndexName, Lokomotive gefundeneLok)
-        {
-            LokKontroll_UpdateAktiveLok(IndexName, gefundeneLok);
-        }
 
 
         #endregion
